Check if filesystem is full

+++ filesystem_check_readonly_mode.sh
@@ -7,14 +7,18 @@ listeMount=`df -PTlk $excludefs | sed -e 1d | awk '{print $7}'`
 hostname=`hostname`
 
 for m in $listeMount; do
-#    echo "Check of $m"
+    space_free=$(df -hP $m | tail -1 | sed 's/ \+/ /g' | cut -d" " -f 5)
+    if  [[ "${space_free%\%}" -eq '100' ]]; then
+        exit 1
+    fi
+
     tmpFile=`mktemp -p "$m" 2>&1`
     rc=$?
     if [[ $rc -ne 0 ]]; then
-        echo "  $hostname ; ERROR ; $tmpFile has an error"
+        echo "  $hostname : ERROR : $tmpFile"
         test -f "$tmpFile" && rm -f "$tmpFile"
     else
-        echo "  $hostname ; INFO ; filesystem $m not in read only mode"
+        echo "  $hostname : INFO : filesystem $m not in read only mode"
         rm -f "$tmpFile"
     fi
 done
